requirements.txt
openai>=1.0
anthropic>=0.39
google-genai>=1.0
rich>=13.0
prompt-toolkit>=3.0
python-dotenv>=1.0
